TypeByRegion Game Mode Initiated
Generated 9 cells depending on screen size. With the "RegionalPositionToCellNumber", we can use this for an array of cluster keys to access.

Next step is setting up the mode and reconfiguring the NoteEffect and CloseInEffect cScripts when changing to this mode, and having the keys appear little earlier before an approach cirlce fades in.