Refactoring
Refactored MapReader.

In the future, we'll have to consider using .json files for the formatting of the playing field.

Start making 3 seperate difficulties of the same song, starting by creating a class that takes the GROUPID of a song to pair with a song's ID to know that it's a difficult of that song.