SongSelectionMenu reads in audio and plays it based on song
You can select a song, and the song will preview itself, as well as show the cover art that is associated with it. Next step that we need to take is having a way to take a record of it for Easy, Normal, Hard, and Insane modes. I feel it's most appropriate to save this record through a .JSON file with the respective difficulty. Not only that, but have an object that hold data on the song difficulty like the highest score on that difficulty, the approach speed, accuracy harshness, total notes, attempts, etc.