Further Improving Song Selection Menu
Improving the existing layout for the Song Selection Menu. You can change difficulty to display, which will update the score and ranking of the song selected in the Song Selection Menu. Different stats will pop up as well when you click on the Next button. 
When changing difficulty, I have it set up where if you click on what difficulty mode you want to be on. There will be further implementation for thus, such as updating the button that sends you to change the difficulty, as well as getting ProjectROFT's mascot in. 0 ~ 0