Fixing Post Processing Issue
I realized that the Post Processing only can render when a canvas is rendered on a specific camera. It can not run if a canvas is marked as a Screen Overlay. So, I had to rearrange some objects and such just to make sure the post processing is running through the whole lifetime of the game.