You can add multiple notes via Pattern Set Stack Value
The issue with the editor was that when adding multiple notes, the one note you have originally placed would be overwritten. By adding a "Stack" mechanic to the editor, you can now "stack" the amount of objects in the same sample position of the track, allowing us to have multiple note objects appearing on the screen during game play.